Fix test-tools.cake

+++ MobileCenterDemoApp/Assets/MobileCenter/Editor/MobileCenterSettingsMakerAndroid.cs
@@ -79,12 +79,17 @@ public class MobileCenterSettingsMakerAndroid
     public void CommitSettings()
     {
         var appId = ApplicationIdHelper.GetApplicationId();
-        var manifestText = File.ReadAllText(MobileCenterManifestPlaceholderPath);
-        if (manifestText.Contains(ManifestAppIdPlaceholder))
+        if (File.Exists(MobileCenterManifestPlaceholderPath))
         {
-            manifestText = manifestText.Replace(ManifestAppIdPlaceholder, appId);
-            File.WriteAllText(MobileCenterManifestPath, manifestText);
+            var manifestText = File.ReadAllText(MobileCenterManifestPlaceholderPath);
+            File.Create(MobileCenterManifestPath).Dispose();
+            if (manifestText.Contains(ManifestAppIdPlaceholder))
+            {
+                manifestText = manifestText.Replace(ManifestAppIdPlaceholder, appId);
+                File.WriteAllText(MobileCenterManifestPath, manifestText);
+            }
         }
+        
         if (File.Exists(MobileCenterResourcesPath))
         {
             File.Delete(MobileCenterResourcesPath);
